test(flow): the instrument that catches this belongs to every binding
The fix was right and its test was in the wrong module. The defect lives
in CoreFlowRuntime, so a Core-local test proved it fixed there — and left
the one thing that can SEE it where no binding inherits it.

That matters more here than usual, because the blindness is structural,
not accidental. ADR-013 §5 makes schemaVersion the optimistic-lock
counter and lets an in-memory store "ignore the field entirely"; every
store a binding hands the TCK does exactly that. So all 13 cases in
Migration pass against a runtime whose migration write leaves the
instance a version behind the row. What was missing was never a case —
the suite is thorough — but something able to see one fail.

So VersionEnforcingSnapshotStore moves to exeris-kernel-tck as a shared
fixture, and the case moves with it as a nested OptimisticLock class. To
let a case hand the engine its own store, createEngine() takes one as a
parameter instead of the binding wiring it privately: a store is bound
when an engine starts, so swapping it afterwards would leave the running
workers writing to the old one.

Proof it is not just relocated but live: reverted against the pre-fix
CoreFlowRuntime, the Community binding runs 30 cases and exactly ONE
fails — the new one. The other 29 certify the bug, which is the finding
restated as a measurement.

The Core-local test is deleted rather than kept alongside. The Community
binding runs the same engine through the provider, so keeping both would
assert the same fact twice while implying two independent checks.

Verified: `mvn clean install` on JDK 25.0.3-tem, no skip flags, full
reactor green, exit 0 (tck 32, core 1516, community 28 + the TCK suite).
